# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ArcadeMatrix RPi Local API
  description: API for controlling and configuring the ArcadeMatrix Raspberry Pi firmware locally over Wi-Fi.
  version: 1.0.0
servers:
  - url: http://arcadematrix.local:8080
    description: Default local mDNS address
tags:
  - name: System
    description: System status and power controls
  - name: Configuration
    description: Global settings and themes
  - name: Media
    description: Fonts, GIFs, and Playlists
  - name: Engines
    description: Direct control of specific display engines

paths:
  /api/system_info:
    get:
      tags:
        - System
      summary: Get system status
      description: Returns memory usage, uptime, cpu load, and temperature.
      responses:
        '200':
          description: System status
          content:
            application/json:
              schema:
                type: object
                properties:
                  cpu_load:
                    type: number
                  ram_used_mb:
                    type: integer
                  ram_total_mb:
                    type: integer
                  ram_percent:
                    type: number
                  disk_free_gb:
                    type: number
                  disk_total_gb:
                    type: number
                  disk_percent:
                    type: number
                  temperature_c:
                    type: number

  /api/settings:
    get:
      tags:
        - Configuration
      summary: Get all settings
      description: Returns the entire device configuration.
      responses:
        '200':
          description: Configuration object
    post:
      tags:
        - Configuration
      summary: Update settings
      description: Partially update settings. Any omitted keys will remain unchanged. Will automatically save to conf.ini.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Settings successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string

  /api/system/power:
    post:
      tags:
        - System
      summary: Toggle LED Matrix Power
      description: Turn the display on or off.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                state:
                  type: boolean
      responses:
        '200':
          description: Success

  /api/system/reboot:
    post:
      tags:
        - System
      summary: Reboot the Raspberry Pi
      responses:
        '200':
          description: Device will reboot

  /api/system/shutdown:
    post:
      tags:
        - System
      summary: Shutdown the Raspberry Pi
      responses:
        '200':
          description: Device will shutdown

  /api/wifi:
    post:
      tags:
        - System
      summary: Reconfigure Wi-Fi
      description: Updates the Wi-Fi credentials via NetworkManager.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ssid:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successfully connected to new network
        '500':
          description: Failed to connect

  /api/mqtt/install:
    post:
      tags:
        - System
      summary: Install remote scripts via SSH
      description: Used to pair the Pi with a retro-gaming console to push game states via MQTT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ip:
                  type: string
      responses:
        '200':
          description: Successfully installed scripts

  /api/fonts:
    get:
      tags:
        - Media
      summary: List available custom fonts
      description: Returns a list of font files found in the /fonts directory.
      responses:
        '200':
          description: Array of font paths
          
  /api/playlists:
    get:
      tags:
        - Media
      summary: Get all GIF playlists
      responses:
        '200':
          description: JSON object mapping playlist categories to arrays of GIF file paths.

  /api/playlists/selected:
    get:
      tags:
        - Media
      summary: Get active GIF playlists
      responses:
        '200':
          description: Selected playlists

  /api/playlists/save:
    post:
      tags:
        - Media
      summary: Save active GIF playlists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                playlists:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Success

  /api/clock:
    post:
      tags:
        - Engines
      summary: Change Clock Theme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                clock_theme:
                  type: integer
      responses:
        '200':
          description: Success

  /api/message:
    post:
      tags:
        - Engines
      summary: Display a Marquee Text Message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                color:
                  type: integer
                size:
                  type: integer
                direction:
                  type: string
                speed:
                  type: integer
                timeoutSeconds:
                  type: integer
      responses:
        '200':
          description: Success

  /api/marquee:
    post:
      tags:
        - Engines
      summary: Stream Image Data
      description: Pushes an image (multipart/form-data) directly to the matrix frame buffer.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Image received and displayed.